NG-2 feat: Add list product endpoint
